requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
djangorestframework==3.15.1
geographiclib==2.0
geopy==2.4.1
gunicorn==21.2.0
idna==3.6
isort==5.13.2
packaging==24.0
psycopg2-binary==2.9.7
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
